% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_config.R
\name{config}
\alias{config}
\alias{has_config}
\alias{valid_config}
\title{Check and validate the user configuration file}
\usage{
has_config()

valid_config(config = "~/config.yml")
}
\arguments{
\item{config}{\code{character(1)} The path to the configuration file (default:
\code{"~/config.yml"})}
}
\value{
\itemize{
\item \code{has_config}: a logical named vector indicating if the user has an
ODBC configuration (\code{.odbc.ini}) or a general configuration (\code{config.yml})
file.
\item \code{valid_config}: a logical value indicating if the user has a valid
\code{config.yml} file.
}
}
\description{
These are utilities to check whether the user has an ODBC
configuration file or a general configuration file. These files are used to
connect via ODBC driver either with an \code{.odbc.ini} file or a general
configuration file \code{config.yml} as used by \code{config} R package. For details,
see \code{\link[=clickhouse_connect]{clickhouse_connect()}}. \code{valid_config} checks if the general
configuration file has the required fields for connecting to the Clickhouse
server.
}
\details{
Either the \code{.odbc.ini} file or the \code{config.yml} file is required to
connect to the Clickhouse server. Either file must be in the user's home
directory.
}
\examples{
has_config()
valid_config()
}
